src/main/java/com/conny/loan/service/TermsServiceImpl.java
package com.conny.loan.service;

import com.conny.loan.domain.Terms;
import com.conny.loan.dto.TermsDTO.Request;
import com.conny.loan.dto.TermsDTO.Response;
import com.conny.loan.repository.TermsRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Service;

@Service
@Slf4j
@RequiredArgsConstructor
public class TermsServiceImpl implements TermsService {

    private final TermsRepository termsRepository;
    private final ModelMapper modelMapper;

    @Override
    public Response create(Request request) {
        Terms terms = modelMapper.map(request, Terms.class);
        Terms created = termsRepository.save(terms);

        return modelMapper.map(created, Response.class);
    }
}
